requirements.txt
bottle
jinja2
gunicorn
facebook-sdk
